создаёт board-controller и переносит в него логику отрисовки карточек
